requirements/prod.txt
## Formerly vendor/python/lib
django-picklefield==0.1.9
django-taggit==0.9.3
pytz==2014.10
httplib2==0.7.3
South==0.8.2

## Formerly vendor/packages
Babel==0.9.6
GitPython==0.1.7
amqp==1.4.6
anyjson==0.3.3
beautifulsoup==3.2.1
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
django-threadedcomments==0.5.2
feedparser==5.1.3
html5lib==0.90
jsonpickle==0.3.1
kombu==3.0.24
lockfile==0.8
logilab-astng==0.20.1
logilab-common==0.50.3
ordereddict==1.1
pyflakes==0.4.0
pyparsing==1.5.5
python-dateutil==1.5
recaptcha-client==1.0.6
six==1.9.0
sqlalchemy==0.6.4
django-constance==0.6
django-statici18n==0.4.5
django-appconf==0.6
urllib3==1.8
elasticsearch==0.4.5
django-allauth==0.17.0
requests-oauthlib==0.4.1
python-openid==2.2.5
oauthlib==0.6.3
django-cacheback==0.9
googleanalytics==0.15.3

## Forked, temporary or otherwise
# 2012-08-02
-e git://github.com/lmorchard/bleach.git@2906726af82ed0063a559124cee610bcd2e5859c#egg=bleach
# 2011-11-23

## Advance version, not yet on pypi
# basket-client==0.3.9
-e git://github.com/mozilla/basket-client.git@1ca127a17e17f5e67c2d9f9a500d85e97460f13d#egg=basket-client
# 2013-08-25 (current?)
-e git+https://github.com/linsomniac/python-memcached@794b02811d8c6fdad304955a4602fb1324270f95#egg=python-memcached
# 2010-11-11 (current?)
-e git://github.com/countergram/pytidylib.git@15f8f734e761485d79556c03266353534c987601#egg=pytidylib
# 2013-11-01 (Current.)
-e git://github.com/mozilla/django-badger.git@1a41701179df11df8ca8ca44898b70b6352623d3#egg=django-badger
# 2015-01-13
-e git+https://github.com/elasticsearch/elasticsearch-dsl-py.git@97774f045c23d74253fe2f8794cb2111f7064285#egg=elasticsearch-dsl-py

## Formerly vendor/src - converted to versions from git refs
python-magic==0.4.6
# -e git://github.com/jsocol/commonware.git@b5544185b2d24adc1eb512735990752400ce9cbd#egg=commonware
commonware==0.4.2
# -e git://github.com/brosner/django-timezones.git@adf2fb3dcb353733bc8f8742388e2e34f17e3d70#egg=django-timezones
django-timezones==0.2
requests==2.3.0
# -e git://github.com/andymckay/django-statsd.git@d77e1c60ae07d9a253e084f11969c2803f1800d3#egg=django-statsd
django-statsd-mozilla==0.3.8.5
# -e git://github.com/erikrose/django-tidings.git@6ee3ac5069dfa76cd85b8fb43def72919731b05e#egg=django-tidings
django-tidings==0.4
# -e git://github.com/getsentry/raven-python.git@edd9b1f4fb0ca85b09c1859a228c54f18d6e7d95#egg=raven-python
raven==3.3.12
# -e git+https://github.com/sehmaschine/django-grappelli.git@95558ba99f656ba9bd0736e93317b280c05f4237#egg=django-grappelli
django-grappelli==2.4.5
# -e git+https://github.com/tomchristie/django-rest-framework.git@195790e60b117eff421eb8f04a9f9f3527e797b8#egg=django-rest-framework
djangorestframework==2.3.8
# -e git+https://github.com/jezdez/django-memcached-hashring.git@e66d2cd9255cb802428872360b7fc7071496c063#egg=django-memcached-hashring
django-memcached-hashring==0.1.2
# -e git+https://github.com/Doist/hash_ring.git@40e7a260eea6387bfa5855457ee85b7939ce6ab8#egg=hash_ring
hash_ring==1.3.1
# -e git://github.com/praekelt/django-recaptcha.git@bbb40d4e5f9ec463ac9015450b87d2f073a0315b#egg=django-recaptcha
django-recaptcha==0.0.9

django-honeypot==0.4.0
bitly_api==0.3
django-ratelimit==0.5.0

## Formerly vendor/src - arbitrary commits not matched to versions
# 2011-12-29
-e git://github.com/django-extensions/django-extensions.git@1610e1e59124071e1d01b3814b6f17ff72b526bb#egg=django-extensions
# 2011-11-13
-e git://github.com/fwenzel/django-mozilla-product-details@5a59a50b2f49f4ff555e87ebecc4fe690e71e7b0#egg=django-mozilla-product-details
# 2010-06-18
-e git://github.com/jbalogh/django-multidb-router.git@25ca0c68929144ed9020d89f2ea1e941c7f9aeca#egg=django-multidb-router
# 2013-04-22
-e git://github.com/jbalogh/jingo.git@327dd865b6c350f02ce40092f2983be19de71303#egg=jingo
# 2010-11-05
-e git://github.com/jsocol/jingo-minify.git@b7405d3f93628190bf83cc43b0ff44e2fbb8b3c0#egg=jingo-minify
# 2011-09-09
-e git://github.com/pcraciunoiu/py-wikimarkup.git@bf2046e42a2f2aaae8df233407efbce803342d04#egg=py-wikimarkup
# 2013-02-07
-e git://github.com/clouserw/tower.git@ce2eba049a5146918fa5239d827bdde9e36eadb7#egg=tower
# 2013-12-18
-e git://github.com/jsocol/django-waffle.git@74290c8b5ec1321a50613a475acceaf843405268#egg=django-waffle
# 2011-05-17
-e git://github.com/semente/django-smuggler.git@f2ca1ca3e1e092b0fb426558b124a1f2b6eed23b#egg=smuggler
# 2011-08-24
-e git://github.com/mozilla/funfactory.git@19b0cc4fd64f8e8f468876951bfe5350caca10d9#egg=funfactory
# 2011-12-08
-e git://github.com/ubernostrum/django-soapbox.git@e6c48818913b5afae0d366740e1ab306a3a28aff#egg=soapbox
# 2013-06-26
-e git://github.com/lmorchard/django-teamwork.git@b7dc818d39c92f733cb933814698cedfbbde3622#egg=django-teamwork
# 2013-02-21
-e git://github.com/simplejson/simplejson.git@062aa98fe19c9536b6ede4c0b84f5f3e6711b768#egg=simplejson

## Obsoleted / no longer needed

# -e git://github.com/mozilla/phpbb3-static.git@61b0f3a6b019a9a32291587d7bf87122ead7fb27#egg=phpbb3-static
# importlib==1.0.2 # included in Python 2.7 and newer for 2.x
# setuptools==0.6c12dev-r80622
